Factor out processDocNode and various macro helpers
